Working on InCity database

+++ Code/app/src/main/java/com/ceid/model/payment_methods/Currency.java
@@ -38,8 +38,8 @@ public class Currency implements Serializable
         return this;
     }
 
-    public BigDecimal getValue(){
-        return this.value;
+    public double getValue(){
+        return this.value.doubleValue();
     }
 
     public String getCurrencySymbol(){
